Use nanoid directly instead of using shortid
Shortid is deprecated and the project recommends using nanoid directly.

Also all published versions of shortid rely on older versions of nanoid
that contain known security issues and should be avoided.

This removes the use of shortid and instead directly uses the latest
version of nanoid that does not contain known security vulnerabilities.